Add files via upload
This API also for the Accept or Reject the order after getting the overall information about the order and product in that order, Seller will accept or reject the order. After Seller accept or reject the order the Status of the Order change pending to the accepted or rejected in the database.